% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{convertH2M}
\alias{convertH2M}
\title{convert human genes to mouse genes.}
\usage{
convertH2M(s)
}
\arguments{
\item{s}{a vector of human gene symbols, like IFIT1}
}
\value{
a vector of mouse gene symbols, like Ifit1
}
\description{
convert human genes to mouse genes.
}
\examples{
convertH2M( c("IFIT1") )
}
