% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.fine.R
\name{DimPlot_my}
\alias{DimPlot_my}
\title{Draw UMAP or tSNE plot with circle backgrounded text}
\usage{
DimPlot_my(
  sce,
  group.by = "seurat_clusters",
  reduction = "umap",
  isRaster = T,
  show.legend = F,
  theme.use = theme_classic,
  par.geom_point = list(size = 0.3),
  par.geom_text = list(size = 4, color = "black"),
  par.circle = list(size = 10, stroke = 1.2, fill = "#FFFFFF66"),
  title = "UMAP plot",
  colSet = NULL
)
}
\arguments{
\item{sce}{Seurat obj}

\item{group.by}{group by}

\item{reduction}{reduction method}

\item{isRaster}{is raster?}

\item{show.legend}{default not show legend}

\item{theme.use}{which theme}

\item{par.geom_point}{geom_point parameters}

\item{par.geom_text}{geom_text parameters}

\item{par.circle}{circle parameters}

\item{title}{main title of the figure}

\item{colSet}{color list}
}
\value{

}
\description{
Draw UMAP or tSNE plot with circle backgrounded text
}
\examples{
DimPlot_my(sce, title="pbmc_3k") #default umap
DimPlot_my(sce, reduction = "tsne", title="pbmc_3k tSNE") #tsne
DimPlot_my(sce, reduction = "pca", title="pbmc_3k PCA") #pca
}
