% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{DimPlot_my2}
\alias{DimPlot_my2}
\title{Draw UMAP, my method 2}
\usage{
DimPlot_my2(scObj, group.by = "seurat_clusters", cols = NULL)
}
\arguments{
\item{scObj}{Seurat obj}

\item{group.by}{group by}
}
\value{

}
\description{
Draw UMAP, my method 2
}
\examples{
DimPlot_2(scObj_known2, "ident")
}
