% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{DoHeatmap_my}
\alias{DoHeatmap_my}
\title{heatmap of average expression, by cluster}
\usage{
DoHeatmap_my(
  sce,
  features,
  slot = "scale.data",
  assay = "RNA",
  main = "",
  group.by = "seurat_clusters",
  rotate = F
)
}
\arguments{
\item{sce}{Seurat obj}

\item{features}{gene list}

\item{slot}{which expression type}

\item{assay}{assay}

\item{main}{title}
}
\value{

}
\description{
heatmap of average expression, by cluster
}
\examples{
DoHeatmap_my(scObj_colon, gene_203307, group.by = 'sample')
DoHeatmap_my(scObj_colon, gene_203307, group.by = 'sample', rotate=T)
}
