% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlot_ByColumnList.One.R
\name{DotPlot_ByColumnList}
\alias{DotPlot_ByColumnList}
\title{facet_grid DotPlot by column, only accept named marker gene list.
v0.3
模仿张泽民实验室的图(https://pubmed.ncbi.nlm.nih.gov/34914499/  Fig.2A)，竖排，分面方框}
\usage{
DotPlot_ByColumnList(
  object,
  features,
  scale = TRUE,
  scale.by = "radius",
  idents = NULL,
  cluster.idents = FALSE,
  base_size = list(),
  cols = NULL,
  direction = 1,
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  scale.min = NA,
  scale.max = NA,
  group.by = NULL,
  split.by = NULL
)
}
\arguments{
\item{object}{Seurat obj}

\item{features}{a list of named marker genes}

\item{scale}{scale?}

\item{scale.by}{scale method}

\item{idents}{idents to use}

\item{cluster.idents}{whether cluster idents}

\item{base_size}{font size}

\item{cols}{colors of gene expression}

\item{direction}{color RColorBrewer direction}

\item{col.min}{min?}

\item{col.max}{max?}

\item{dot.min}{dot size min}

\item{dot.scale}{dot max size}

\item{scale.min}{min?}

\item{scale.max}{max?}

\item{group.by}{group by}

\item{split.by}{split by}
}
\value{

}
\description{
facet_grid DotPlot by column, only accept named marker gene list.
v0.3
模仿张泽民实验室的图(https://pubmed.ncbi.nlm.nih.gov/34914499/  Fig.2A)，竖排，分面方框
}
\examples{
DotPlot_ByColumnList(scObj_colon, features = common_markers)
}
