% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlot_my.One.R
\name{DotPlot_my}
\alias{DotPlot_my}
\title{My enhanced DotPlot
version:0.1}
\usage{
DotPlot_my(
  object,
  assay = NULL,
  features,
  cols = colorRampPalette(colors = c("lightyellow", "orangered", "red4"), bias = 0.5,
    interpolate = c("spline"))(100),
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  idents = NULL,
  group.by = NULL,
  split.by = NULL,
  cluster.idents = FALSE,
  scale = TRUE,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA,
  circle_size = 6,
  base_size = list(x = 14, y = 14, top = 14, legend = 14),
  rel_widths = c(1, 8),
  color_list = NULL
)
}
\arguments{
\item{object}{Seurat obj}

\item{assay}{which assay}

\item{features}{gene list}

\item{cols}{colors for expression}

\item{col.min}{min?}

\item{col.max}{max?}

\item{dot.min}{dot min size}

\item{dot.scale}{dot scale method}

\item{idents}{which idents to show}

\item{group.by}{group by}

\item{split.by}{split by}

\item{cluster.idents}{is cluster idents?}

\item{scale}{scale?}

\item{scale.by}{scale by?}

\item{scale.min}{min?}

\item{scale.max}{max?}

\item{base_size}{font size list}

\item{rel_widths}{relative widths}

\item{color_list}{color list for each cluster}
}
\value{

}
\description{
My enhanced DotPlot
version:0.1
}
\examples{
DotPlot_my(sce, features = c("CD3D", "CD79A"))
DotPlot_my(sce, features = c("CD3D", "CD79A"), cols = c("white", "red2"), rel_widths = c(1,4))
colors=c("#96C3D8", "#5F9BBE", "#F5B375", "#C0937E", "#67A59B", "#A5D38F", "#4A9D47", "#F19294", "#E45D61", "#3377A9",
"#BDA7CB", "#684797", "#8D75AF", "#CD9C9B", "#D62E2D", "#DA8F6F", "#F47D2F")
DotPlot_my(pbmc, features = split(top11$gene, top11$cluster), cols = c("white", "red3"), color_list=colors, rel_widths=c(1,45))
DotPlot_my(scObj_colon, features = common_markers, color_list = colors,
          rel_widths = c(1,35), base_size = list(legend=10, x=12))
}
