% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlot_my2.One.R
\name{DotPlot_my2}
\alias{DotPlot_my2}
\title{My enhanced DotPlot}
\usage{
DotPlot_my2(
  object,
  features,
  assay = NULL,
  cols = colorRampPalette(colors = c("lightyellow", "orangered", "red4"), bias = 0.5,
    interpolate = c("spline"))(100),
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  idents = NULL,
  group.by = NULL,
  split.by = NULL,
  cluster.idents = FALSE,
  scale = TRUE,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA,
  cluster.features = T,
  quick = T,
  circle_size = 6,
  base_size = list(x = 14, y = 14, top = 14, legend = 14),
  rel_widths = c(1, 8),
  color_list = NULL
)
}
\arguments{
\item{object}{Seurat obj}

\item{features}{gene symbol}

\item{assay}{which assay}

\item{cols}{colors for expression}

\item{col.min}{min?}

\item{col.max}{max?}

\item{dot.min}{dot min size}

\item{dot.scale}{dot scale method}

\item{idents}{which ident to show}

\item{group.by}{group by}

\item{split.by}{split by}

\item{cluster.idents}{whether cluster idents?}

\item{scale}{scale?}

\item{scale.by}{scale by?}

\item{scale.min}{min?}

\item{scale.max}{max?}

\item{cluster.features}{whether cluster idents?}

\item{quick}{return normal ggplot2 obj as Seurat}

\item{circle_size}{circle size left of y axis}

\item{base_size}{list of font size}

\item{rel_widths}{relative widths, only effective when quick =F}

\item{color_list}{color list for each cluster}
}
\value{

}
\description{
v0.2 can cluster genes
}
\examples{
DotPlot_my2(sce, features = c("CD3D", "CD79A"))
}
