% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{DoVlnPlot_stack}
\alias{DoVlnPlot_stack}
\title{Plot stack violin plot from Seurat object}
\usage{
DoVlnPlot_stack(
  object,
  features,
  cols = NULL,
  group.by = NULL,
  type = "h",
  fill_by = "cluster"
)
}
\arguments{
\item{object}{Seurat obj}

\item{features}{gene symbols}

\item{cols}{color list}

\item{group.by}{group by}

\item{type}{which type, h or v}

\item{fill_by}{like group by}
}
\value{

}
\description{
Plot stack violin plot from Seurat object
}
\examples{
genes2=c("Cd79a", "Igha", "Cd3d", "Itgax", "Ly6g");
colors2=c("red", "blue", "red3", "orange", "purple");
DoVlnPlot_stack(scObj_known2, features = genes2, cols =colors2,
  group.by = "ident", fill_by = "gene", type="v")
}
