% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{Feature_scale_PCA_all}
\alias{Feature_scale_PCA_all}
\title{shortcut: HVG, Scale(all), PCA}
\usage{
Feature_scale_PCA_all(object, nfeatures)
}
\arguments{
\item{object}{Seurat obj}

\item{nfeatures}{number of HVG}
}
\value{

}
\description{
shortcut: HVG, Scale(all), PCA
}
\examples{
Feature_scale_PCA_all(sce, 3000)
}
