% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{Feature_scale_PCA_variable}
\alias{Feature_scale_PCA_variable}
\title{shortcut: HVG, Scale(HVG), PCA}
\usage{
Feature_scale_PCA_variable(object, nfeatures)
}
\arguments{
\item{object}{Seurat obj}

\item{nfeatures}{numbre of HVG}
}
\value{

}
\description{
shortcut: HVG, Scale(HVG), PCA
}
\examples{
Feature_scale_PCA_variable(sce, 3000)
}
