% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{FindMarkersDIY}
\alias{FindMarkersDIY}
\title{FindMarkersDIY, quick version}
\usage{
FindMarkersDIY(object = object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
FindMarkersDIY, quick version
}
\examples{
FindMarkersDIY(sce)
}
