% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.fine.R
\name{getGeneScoreValue}
\alias{getGeneScoreValue}
\title{Calc gene score}
\usage{
getGeneScoreValue(sce, gene_list, group.by = NULL)
}
\arguments{
\item{sce}{Seurat obj}

\item{gene_list}{gene list}

\item{group.by}{group by}
}
\value{

}
\description{
Calc gene score
}
\examples{
getGeneScoreValue(sce, gene_list)
}
