% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{getLongDf}
\alias{getLongDf}
\title{get 3col-df (exp value/ gene symbol /  cluster id) for stack violin plot}
\usage{
getLongDf(scRNA, genenames, group.by = "seurat_clusters")
}
\arguments{
\item{scRNA}{Seurat obj}

\item{genenames}{gene symbol list}

\item{group.by}{use which cluster cretia}
}
\value{

}
\description{
get 3col-df (exp value/ gene symbol /  cluster id) for stack violin plot
}
\examples{
df2= getLongDf(scObj_known2, c("Cd3e", "Cd19"), group.by='sample')
}
