% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_Seurat_related.R
\name{getSankeyData}
\alias{getSankeyData}
\title{Title: Make Sankey plot data from a table(paraLeft, paraRight)
ref: https://observablehq.com/ at d3/sankey
version: 2.0 define flow color as source//done
v2.1 supply 3 color method: static, source, target//done}
\usage{
getSankeyData(
  tbl1,
  prefix_left = "",
  prefix_right = "",
  colorMethod = "source",
  omitR = NULL,
  omitC = NULL,
  rm.na = T
)
}
\arguments{
\item{tbl1}{a table produced by table()}

\item{prefix_left}{add a prefix to left nodes}

\item{prefix_right}{add a prefix to right nodes}

\item{colorMethod}{"static", "source", "target"}

\item{omitR}{omitR rownames}

\item{omitC}{omitC column names}

\item{omit}{omit which rows/paraLeft}
}
\value{

}
\description{
Title: Make Sankey plot data from a table(paraLeft, paraRight)
ref: https://observablehq.com/ at d3/sankey
version: 2.0 define flow color as source//done
v2.1 supply 3 color method: static, source, target//done
}
\examples{
getSankeyData( table(mtcars$gear, mtcars$carb))
}
