% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{Neighbor_cluster_umap_tsne}
\alias{Neighbor_cluster_umap_tsne}
\title{shortcut: cell Cluster, UMAP}
\usage{
Neighbor_cluster_umap_tsne(object, dims, resolution)
}
\arguments{
\item{object}{Seurat obj}

\item{dims}{dimentions to use of PC}

\item{resolution}{a number between 0.1-2, sometimes evern 50, 200}
}
\value{

}
\description{
shortcut: cell Cluster, UMAP
}
\examples{
Neighbor_cluster_umap_tsne(sce, 1:30, 0.8)
}
