% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{nFeature.hist}
\alias{nFeature.hist}
\title{hist plot of nFeature in each cluster}
\usage{
nFeature.hist(object, cluster, xlim = c(0, 500))
}
\arguments{
\item{object}{Seurat obj}

\item{cluster}{which cluster}

\item{xlim}{limits of x axis}
}
\value{

}
\description{
hist plot of nFeature in each cluster
}
\examples{
nFeature.hist(sce, 2, c(0,200))
}
