% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{nFeatureScatter}
\alias{nFeatureScatter}
\title{2D FACS like scatter plot}
\usage{
nFeatureScatter(
  object,
  feature1,
  feature2,
  cells = NULL,
  group.by = NULL,
  cols = NULL,
  pt.size = 1,
  shape.by = NULL,
  span = NULL,
  smooth = FALSE,
  combine = TRUE,
  slot = "data",
  plot.cor = TRUE,
  raster = NULL
)
}
\arguments{
\item{object}{Seurat ojb}

\item{feature1}{gene symbol 1}

\item{feature2}{gene symbol 2}

\item{cells}{cid}

\item{group.by}{group by}

\item{cols}{colors}

\item{pt.size}{dot size}

\item{shape.by}{shape by}

\item{span}{span?}

\item{smooth}{smooth?}

\item{combine}{combine?}

\item{slot}{slot}

\item{plot.cor}{plot.correlation?}

\item{raster}{is raster}
}
\value{

}
\description{
2D FACS like scatter plot
}
\examples{
nFeatureScatter(sce, "CD3D", "CD4")
}
