% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.fine.R
\name{plotGeneScore}
\alias{plotGeneScore}
\title{plot violin plot of gene score among groups(ggplot2)}
\usage{
plotGeneScore(sce, gene_list, group.by = NULL)
}
\arguments{
\item{sce}{Seurat obj}

\item{gene_list}{gene list}

\item{group.by}{group by}
}
\value{

}
\description{
plot violin plot of gene score among groups(ggplot2)
}
\examples{
plotGeneScore(sce, gene_list)
}
