% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlot_ByColumnList.One.R
\name{print.GEB_lite}
\alias{print.GEB_lite}
\title{回车直接画 grid 图形}
\usage{
\method{print}{GEB_lite}(x, newpage = TRUE)
}
\arguments{
\item{x}{an Grob object with class GEB_lite}

\item{newpage}{whether newpage?}
}
\value{

}
\description{
回车直接画 grid 图形
}
\examples{
print(sth)
}
