% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{Q_cellNumber}
\alias{Q_cellNumber}
\title{每个亚群中的细胞数量饼图}
\usage{
Q_cellNumber(scObj, group.by = "seurat_clusters", colors = NULL)
}
\arguments{
\item{scObj}{Seurat obj}

\item{group.by}{group by}

\item{colors}{color list}
}
\value{

}
\description{
每个亚群中的细胞数量饼图
}
\examples{
Q_cellNumber(sce)
Q_cellNumber(scObj_known2, colors=allcolour)
Q_cellNumber(scObj_known2, "ident", colors=allcolour)

}
