% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{Q_sankey}
\alias{Q_sankey}
\title{Title: 快速桑基图，几种分类方法的来源和去向}
\usage{
Q_sankey(scObj0, col_list = NULL, downsample = NULL, color_res = 1)
}
\arguments{
\item{scObj0}{Seurat obj}

\item{col_list}{column list of meta data}

\item{downsample}{sample size of each cluster, to get a small subset}

\item{color_res}{color by resolution, default 1}
}
\value{

}
\description{
Title: 快速桑基图，几种分类方法的来源和去向
}
\examples{
Q_sankey(pbmc)
Q_sankey(pbmc, c(5, 7))
}
