% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{QC_nGene_nUMI_expr}
\alias{QC_nGene_nUMI_expr}
\title{plot gene expression level with color, on nGene x nUMI plot}
\usage{
QC_nGene_nUMI_expr(scObj, gene_symbol = NULL, keyword = "")
}
\arguments{
\item{scObj}{Seurat obj}

\item{gene_symbol}{gene symbol}

\item{keyword}{sample info on title}
}
\value{

}
\description{
plot gene expression level with color, on nGene x nUMI plot
}
\examples{
QC_nGene_nUMI_expr(sce)
QC_nGene_nUMI_expr(sce, "CD4")
}
