% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_Seurat_related.R
\name{sankeyPlot}
\alias{sankeyPlot}
\title{Sankey plot from sankeyDt, produced by getSankeyData()}
\usage{
sankeyPlot(sankeyDt, colors = NULL)
}
\arguments{
\item{sankeyDt}{dataset produced by getSankeyData}

\item{colors}{a color list for sankeyDt$node$node}
}
\value{

}
\description{
version: 2.0 define flow color as source//done
depend getSankeyData()
}
\examples{
getSankeyData( table(mtcars$gear, mtcars$carb))
sankeyDt2=getSankeyData(table(mtcars$gear, mtcars$carb),"Gear_", "Carb_")
sankeyPlot(sankeyDt2 )
sankeyPlot( getSankeyData( table(mtcars$gear, mtcars$carb), "Gear_", "Carb_", colorMethod="source") )
sankeyPlot( getSankeyData( table(mtcars$gear, mtcars$carb), "Gear_", "Carb_", colorMethod="target") )
sankeyPlot( getSankeyData( table(mtcars$gear, mtcars$carb), "Gear_", "Carb_", colorMethod="static") )

library(RColorBrewer);
mypalette_1=brewer.pal(8,"Set1")
sankeyPlot(sankeyDt2, colors=mypalette_1 )
}
