% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{showCluster}
\alias{showCluster}
\title{Show the position of a cluster on umap}
\usage{
showCluster(scObj, clusterId, slot = "seurat_clusters", color = "darkred")
}
\arguments{
\item{scObj}{Seurat obj}

\item{clusterId}{which cluster}

\item{slot}{default seurat_clusters}

\item{color}{highlight color, default darkred}
}
\value{

}
\description{
Show the position of a cluster on umap
}
\examples{
showCluster(scObj, 17)
showCluster(scObj, "BL1Y", slot="sample")
showCluster(scObj, "BL1Y", slot="sample", color="purple")
}
