% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{SingleCorPlot}
\alias{SingleCorPlot}
\title{SingleCorPlot}
\usage{
SingleCorPlot(
  data,
  col.by = NULL,
  cols = NULL,
  pt.size = NULL,
  smooth = FALSE,
  rows.highlight = NULL,
  legend.title = NULL,
  na.value = "grey50",
  span = NULL,
  raster = NULL,
  plot.cor = TRUE
)
}
\arguments{
\item{data}{matrix?}

\item{col.by}{color by?}

\item{cols}{colors?}

\item{pt.size}{dot size}

\item{smooth}{smooth?}

\item{rows.highlight}{which row?}

\item{legend.title}{legend title}

\item{na.value}{value of NA}

\item{span}{span?}

\item{raster}{is raster?}

\item{plot.cor}{plot correlation}
}
\value{

}
\description{
SingleCorPlot
}
\examples{
SingleCorPlot(sce)
}
