% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{table2barplot}
\alias{table2barplot}
\title{draw barplot, from a data.frame generated by table(para1, para2), colored by 1st parameter of table()}
\usage{
table2barplot(
  tbl1,
  colors = NULL,
  levels = NULL,
  scale = T,
  title = "",
  omit = NULL,
  xlab = "",
  ylab = "",
  legendTitle = NULL
)
}
\arguments{
\item{tbl1}{data.frame, draw by each column}

\item{colors}{colors of each row(default NULL, auto-color)}

\item{scale}{whether scale to 1 or not(default T)}

\item{title}{the main title of the figure,(is main in the function)}

\item{omit}{whether to omit some columns}

\item{legendY}{the position y of legend, adjust as needed, default -0.25}

\item{...}{}
}
\description{
v2.1
v2.2 第一参数的空值跳过
}
\examples{
table2barplot(table(sce$seurat_clusters, sce$sample) )
}
