% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{VlnPlot_stack}
\alias{VlnPlot_stack}
\title{Plot stacked violin plot using 3col-df(value/ gene/  cluster)}
\usage{
VlnPlot_stack(df1, type = "h", colors = NULL, fill_by = "cluster")
}
\arguments{
\item{df1}{df of 3 columns}

\item{type}{h or v}

\item{colors}{color list}

\item{fill_by}{like group by}
}
\value{

}
\description{
Plot stacked violin plot using 3col-df(value/ gene/  cluster)
}
\examples{
VlnPlot_stack(df1, "h")
}
