% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.fine.R
\name{FeaturePlot_my}
\alias{FeaturePlot_my}
\title{enhenced FeaturePlot, with one unified legend color-bar.}
\usage{
FeaturePlot_my(
  sce,
  features,
  ncol = 5,
  cut_dims = c(-1, 1.5),
  colors = brewer.pal(9, "YlOrRd")
)
}
\arguments{
\item{sce}{Seurat obj}

\item{features}{gene symbol list}

\item{ncol}{num of columns}

\item{cut_dims}{the min,max of scaled data}

\item{colors}{color list, default will be ok}
}
\value{

}
\description{
color scheme: red like brown, z-score used.
}
\examples{
FeaturePlot_my(scObj_colon, features =c("Itgam", "Ly6g", "Cxcr2", "Cxcr4", "Cxcl10"), ncol=5 )
}
