% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_related.R
\name{meanNFeatureByCluster}
\alias{meanNFeatureByCluster}
\title{select clusters by mean
v2.0 order output}
\usage{
meanNFeatureByCluster(object, nth, draw = T)
}
\arguments{
\item{object}{Seurat obj}

\item{nth}{a number}

\item{draw}{wether output text of draw picture}
}
\value{

}
\description{
select clusters by mean
v2.0 order output
}
\examples{
rs=meanNFeatureByCluster(sce, 3)
}
